refactor user registration logic; improve error handling for file uploads; update user model to allow non-unique full names; fix multer storage path for Windows compatibility; enhance cloudinary upload error handling

+++ src/utils/cloudinary.js
@@ -15,11 +15,17 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: 'auto'
         })
         //file has been uploaded successfully
-        console.log('file is uploaded on cloudinary', response.url);
+        // console.log('file is uploaded on cloudinary', response.url);
+        console.log(response);
+        
+        // fs.unlinkSync(localFilePath);
         return response;
         
     } catch (error) {
-        fs.unlinkSync(localFilePath); //remove the locally saved temp file as the upload operation got failed
+        console.log(error)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath); //remove the locally saved temp file as the upload operation got failed
+        }
     }
 }
